pyproject.toml
[tool.poetry]
name = "main"
version = "0.1.0"
description = "hello world"
authors = ["jshimazu"]

[tool.poetry.dependencies]
python = "^3.8"
things-py = "0.0.14"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"